Separate UI and Logic for dashboard
